+++ iconsdk/icon_service.py
@@ -12,6 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from typing import Union
 
 from iconsdk.builder.call_builder import Call
 from iconsdk.builder.transaction_builder import Transaction
@@ -43,7 +44,7 @@ class IconService:
     def __init__(self, provider: Provider):
         self.__provider = provider
 
-    def get_block(self, value):
+    def get_block(self, value: Union[int, str], full_response: bool = False) -> dict:
         """
         If param is height,
             1. Returns block information by block height
@@ -61,102 +62,128 @@ class IconService:
             Integer of a block height
             or hash of a block prefixed with '0x'
             or `latest`
+        :param full_response:
+            Boolean to check whether get naive dict or refined data from server
         :return result: Block data
         """
         # by height
         if is_block_height(value):
             params = {'height': add_0x_prefix(hex(value))}
-            result = self.__provider.make_request('icx_getBlockByHeight', params)
+            result = self.__provider.make_request('icx_getBlockByHeight', params, full_response)
         # by hash
         elif is_hex_block_hash(value):
             params = {'hash': value}
-            result = self.__provider.make_request('icx_getBlockByHash', params)
+            result = self.__provider.make_request('icx_getBlockByHash', params, full_response)
         # by value
         elif is_predefined_block_value(value):
-            result = self.__provider.make_request('icx_getLastBlock')
+            result = self.__provider.make_request('icx_getLastBlock', full_response=full_response)
         else:
             raise DataTypeException("It's unrecognized block reference:{0!r}.".format(value))
 
-        convert_block(result)
+        if not full_response:
+            convert_block(result)
+
         return result
 
-    def get_total_supply(self):
+    def get_total_supply(self, full_response: bool = False) -> Union[dict, int]:
         """
         Returns total ICX coin supply that has been issued
         Delegates to icx_getTotalSupply RPC method
 
+        :param full_response: Boolean to check whether get naive dict or refined data from server
         :return: Total number of ICX coins issued
         """
-        result = self.__provider.make_request('icx_getTotalSupply')
-        return int(remove_0x_prefix(result), 16)
 
-    def get_balance(self, address: str):
+        result = self.__provider.make_request('icx_getTotalSupply', full_response=full_response)
+
+        if full_response:
+            return result
+        else:
+            return int(remove_0x_prefix(result), 16)
+
+    def get_balance(self, address: str, full_response: bool = False) -> Union[dict, int]:
         """
         Returns the ICX balance of the given EOA or SCORE.
         Delegates to icx_getBalance RPC method.
 
         :param address: An address of EOA or SCORE. type(str)
+        :param full_response: Boolean to check whether get naive dict or refined data from server
         :return: Number of ICX coins
         """
         if is_score_address(address) or is_wallet_address(address):
             params = {'address': address}
-            result = self.__provider.make_request('icx_getBalance', params)
-            return int(remove_0x_prefix(result), 16)
+
+            result = self.__provider.make_request('icx_getBalance', params, full_response)
+
+            if full_response:
+                return result
+            else:
+                return int(remove_0x_prefix(result), 16)
         else:
             raise AddressException("Address is wrong.")
 
-    def get_score_api(self, address: str):
+    def get_score_api(self, address: str, full_response: bool = False) -> Union[dict, list]:
         """
         Returns SCORE's external API list.
         Delegates to icx_getScoreApi RPC method.
 
         :param address: A SCORE address to be examined
+        :param full_response: Boolean to check whether get naive dict or refined data from server
         :return: A list of API methods of the SCORE and its information
         """
         if not is_score_address(address):
             raise AddressException("SCORE Address is wrong.")
 
         params = {'address': address}
-        return self.__provider.make_request('icx_getScoreApi', params)
+        return self.__provider.make_request('icx_getScoreApi', params, full_response)
 
-    def get_transaction_result(self, tx_hash: str):
+    def get_transaction_result(self, tx_hash: str, full_response: bool = False) -> dict:
         """
         Returns the transaction result requested by transaction hash.
         Delegates to icx_getTransactionResult RPC method.
 
         :param tx_hash: Hash of a transaction prefixed with '0x'
+        :param full_response: Boolean to check whether get naive dict or refined data from server
         :return A transaction result object
         """
         if not is_T_HASH(tx_hash):
             raise DataTypeException("This hash value is unrecognized.")
 
         params = {'txHash': tx_hash}
-        result = self.__provider.make_request('icx_getTransactionResult', params)
-        convert_transaction_result(result)
+        result = self.__provider.make_request('icx_getTransactionResult', params, full_response)
+
+        if not full_response:
+            convert_transaction_result(result)
+
         return result
 
-    def get_transaction(self, tx_hash: str):
+    def get_transaction(self, tx_hash: str, full_response: bool = False) -> dict:
         """
         Returns the transaction information requested by transaction hash.
         Delegates to icx_getTransactionByHash RPC method.
 
         :param tx_hash: Transaction hash prefixed with '0x'
+        :param full_response: Boolean to check whether get naive dict or refined data from server
         :return: Information about a transaction
         """
         if not is_T_HASH(tx_hash):
             raise DataTypeException("This hash value is unrecognized.")
 
         params = {'txHash': tx_hash}
-        result = self.__provider.make_request('icx_getTransactionByHash', params)
-        convert_transaction(result)
+        result = self.__provider.make_request('icx_getTransactionByHash', params, full_response)
+
+        if not full_response:
+            convert_transaction(result)
+
         return result
 
-    def call(self, call: object):
+    def call(self, call: object, full_response: bool = False) -> Union[dict, str]:
         """
         Calls SCORE's external function which is read-only without creating a transaction on Loopchain.
         Delegates to icx_call RPC method.
 
         :param call: Call object made by CallBuilder
+        :param full_response: Boolean to check whether get naive dict or refined data from server
         :return: Values returned by the executed SCORE function
         """
         if not isinstance(call, Call):
@@ -176,18 +203,19 @@ class IconService:
         if isinstance(call.params, dict):
             params["data"]["params"] = call.params
 
-        return self.__provider.make_request('icx_call', params)
+        return self.__provider.make_request('icx_call', params, full_response)
 
-    def send_transaction(self, signed_transaction: SignedTransaction):
+    def send_transaction(self, signed_transaction: SignedTransaction, full_response: bool = False) -> Union[dict, str]:
         """
         Sends the transaction.
         Delegates to icx_sendTransaction RPC method.
 
         :param signed_transaction: A signed transaction object
+        :param full_response: Boolean to check whether get naive dict or refined data from server
         :return: Transaction hash prefixed with '0x'
         """
         params = signed_transaction.signed_transaction_dict
-        return self.__provider.make_request('icx_sendTransaction', params)
+        return self.__provider.make_request('icx_sendTransaction', params, full_response)
 
     def estimate_step(self, transaction: Transaction) -> int:
         """