+++ iconsdk/libs/serializer.py
@@ -15,6 +15,7 @@
 
 from copy import deepcopy
 from hashlib import sha3_256
+from typing import Optional
 
 translator = str.maketrans({
     "\\": "\\\\",
@@ -26,9 +27,9 @@ translator = str.maketrans({
 })
 
 
-def __make_params_serialized(json_data: dict):
+def __make_params_serialized(json_data: dict) -> str:
 
-    def encode(data):
+    def encode(data) -> str:
         if isinstance(data, dict):
             return encode_dict(data)
         elif isinstance(data, list):
@@ -36,24 +37,24 @@ def __make_params_serialized(json_data: dict):
         else:
             return escape(data)
 
-    def encode_dict(data: dict):
+    def encode_dict(data: dict) -> str:
         result = ".".join(_encode_dict(data))
         return "{" + result + "}"
 
-    def _encode_dict(data: dict):
+    def _encode_dict(data: dict) -> list:
         for key in sorted(data.keys()):
             yield key
             yield encode(data[key])
 
-    def encode_list(data: list):
+    def encode_list(data: list) -> str:
         result = ".".join(_encode_list(data))
         return f"[" + result + "]"
 
-    def _encode_list(data: list):
+    def _encode_list(data: list) -> list:
         for item in data:
             yield encode(item)
 
-    def escape(data):
+    def escape(data) -> str:
         if data is None:
             return "\\0"
 
@@ -85,7 +86,7 @@ def serialize(params: dict) -> bytes:
     return f"icx_sendTransaction.{partial_serialized_params}".encode()
 
 
-def generate_message(params: dict):
+def generate_message(params: dict) -> str:
     """
     Generates transaction's message hash from params in request for transaction.
 
@@ -96,14 +97,14 @@ def generate_message(params: dict):
     return sha3_256(bytes_message_hash).hexdigest()
 
 
-def __get_key_name_for_tx_hash(params):
+def __get_key_name_for_tx_hash(params: dict) -> Optional[str]:
     if __get_tx_version(params) == hex(2):
         return "tx_hash"
     else:
         return None
 
 
-def __get_tx_version(params):
+def __get_tx_version(params: dict) -> str:
     if 'version' not in params:
         return hex(2)
     else: