+++ iconsdk/providers/http_provider.py
@@ -13,15 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Union
 import json
 import os
 from json.decoder import JSONDecodeError
-from logging import getLogger
 from urllib.parse import urlparse
 
 import requests
 from multipledispatch import dispatch
 
+from iconsdk import logger
 from iconsdk.exception import JSONRPCException, URLException
 from iconsdk.providers.config_api_path import CONFIG_API_PATH
 from iconsdk.providers.provider import Provider
@@ -34,13 +35,8 @@ class HTTPProvider(Provider):
     For local development this would be something like 'http://localhost:9000'.
     """
 
-    _logger = getLogger("HTTPProvider")
-
-    # No need to use logging, remove the line.
-    # set_logger(logger, 'DEBUG')
-
     @staticmethod
-    def _validate_base_domain_url(base_domain_url) -> bool:
+    def _validate_base_domain_url(base_domain_url: str) -> bool:
         """
         Validates if input base domain url is valid or not.
 
@@ -67,12 +63,16 @@ class HTTPProvider(Provider):
         :param request_kwargs: kwargs for setting to head of request
         """
         if not self._validate_base_domain_url(base_domain_url):
+            logger.error(f"While setting HTTPProvider, raised URLException because the URL is invalid. "
+                         f"URL: {base_domain_url}")
             raise URLException("Invalid base domain URL. "
                                "Valid base domain URL format is as like <scheme>://<host>:<port>.")
         self._full_path_url = None
         self._base_domain_url = base_domain_url
         self._version = version
         self._request_kwargs = request_kwargs or {}
+        logger.info(f"Set HTTPProvider. "
+                    f"Base domain URL: {base_domain_url}, Version: {version}, Request kwargs: {self._request_kwargs}")
 
     @dispatch(str, dict=None)
     def __init__(self, full_path_url: str, request_kwargs: dict = None):
@@ -85,19 +85,21 @@ class HTTPProvider(Provider):
         """
         self._full_path_url = full_path_url
         self._request_kwargs = request_kwargs or {}
+        logger.info(f"Set HTTPProvider. "
+                    f"Full path URL: {full_path_url}, Request kwargs: {self._request_kwargs}")
 
     def __str__(self):
         return "RPC connection {0}".format(self._base_domain_url)
 
     @to_dict
-    def _get_request_kwargs(self):
+    def _get_request_kwargs(self) -> dict:
         if 'headers' not in self._request_kwargs:
             yield 'headers', {'Content-Type': 'application/json'}
         for key, value in self._request_kwargs.items():
             yield key, value
 
     @staticmethod
-    def _make_post_request(full_path_url, data, **kwargs):
+    def _make_post_request(full_path_url: str, data: dict, **kwargs) -> requests.Response:
         kwargs.setdefault('timeout', 10)
         with requests.Session() as session:
             response = session.post(url=full_path_url, data=json.dumps(data), **kwargs)
@@ -112,7 +114,7 @@ class HTTPProvider(Provider):
         url_components = url_components._replace(path=path)
         return url_components.geturl()
 
-    def make_request(self, method, params=None):
+    def make_request(self, method: str, params=None, full_response: bool = False) -> Union[str, list, dict]:
         rpc_dict = {
             'jsonrpc': '2.0',
             'method': method,
@@ -124,14 +126,18 @@ class HTTPProvider(Provider):
 
         full_path_url = self._full_path_url if self._full_path_url else self._get_full_path_url(method)
         response = self._make_post_request(full_path_url, rpc_dict, **self._get_request_kwargs())
+        custom_response = self._return_custom_response(response, full_response)
 
-        # self._logger.debug("request HTTP\nURI: %s\nMethod: %s\nData: %s", full_path_url, method, rpc_dict)
-        # self._logger.debug("response HTTP\nResponse:%s", response)
+        logger.debug(f"Request: {rpc_dict}")
+        logger.debug(f"Response: {custom_response}")
 
-        return self._return_customed_response(response)
+        return custom_response
 
     @staticmethod
-    def _return_customed_response(response):
+    def _return_custom_response(response: requests.Response, full_response: bool = False) -> Union[str, list, dict]:
+        if full_response:
+            return json.loads(response.content)
+
         if response.ok:
             content_as_dict = json.loads(response.content)
             return content_as_dict["result"]
@@ -139,13 +145,17 @@ class HTTPProvider(Provider):
             try:
                 content_as_dict = json.loads(response.content)
             except (JSONDecodeError, KeyError):
+                logger.exception(f"Raised URLException while returning the custom response. "
+                                 f"Response content: {response.content.decode('utf-8')}")
                 raise URLException(response.content.decode("utf-8"))
             else:
+                logger.error(f"Raised JSONRPCException while returning the custom response. "
+                             f"Error message: {content_as_dict['error']}")
                 raise JSONRPCException(content_as_dict["error"])
 
-    def is_connected(self):
+    def is_connected(self) -> bool:
         try:
-            # self._logger.debug("Connected")
+            logger.debug("Connected")
             last_block = self.make_request('icx_getLastBlock', [])
         except (IOError, URLException, JSONRPCException):
             return False