+++ iconsdk/utils/__init__.py
@@ -13,10 +13,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from os import path
 from functools import wraps
 from logging import StreamHandler, Formatter
+from os import path
 from time import time
+
+from iconsdk import logger
 from iconsdk.utils.hexadecimal import add_0x_prefix
 
 
@@ -40,16 +42,24 @@ def apply_to_return_value(callback):
             return callback(fn(*args, **kwargs))
 
         return inner
+
     return outer
 
 
 to_dict = apply_to_return_value(dict)
 
 
-def set_logger(logger, level):
-    handler = StreamHandler()
-    formatter = Formatter(
-        '%(asctime)s %(name)-12s %(levelname)-8s %(message)s')
+def set_logger(level, handler=StreamHandler(),
+               format: str = '%(asctime)s %(name)-12s %(levelname)-5s '
+                             '%(filename)-12s %(lineno)-4s %(funcName)-12s %(message)s') -> None:
+    """ Set logger by setting level and handler
+
+    :param level: the logging level of this logger. The level must be an int or a str.
+    :param handler: the specified handler to be added to this logger
+    :param format: the specified format strings which initialize the formatter with.
+    :return: None
+    """
+    formatter = Formatter(format)
     handler.setFormatter(formatter)
     logger.addHandler(handler)
     logger.setLevel(level)
@@ -58,5 +68,3 @@ def set_logger(logger, level):
 def get_timestamp():
     """Get epoch time"""
     return add_0x_prefix(hex(int(time() * 10 ** 6)))
-
-